api/requirements.txt
python-dotenv
groq
uvicorn
fastapi[standard]